begin test sweet for server side testing. running npm test cannot find test files matching pattern: test

+++ db/migrations/20170906200234_initial.js
@@ -3,8 +3,8 @@ exports.up = (knex, Promise) => Promise.all([
   knex.schema.createTable('location', (table) => {
     table.increments('id').primary();
     table.string('name');
-    table.decimal('latitude');
-    table.decimal('longitude');
+    table.float('latitude');
+    table.float('longitude');
     table.string('phone_number');
     table.string('website_url');
     table.timestamps(true, true);
